+++ src/test/java/com/braintreegateway/integrationtest/DocumentUploadIT.java
@@ -95,6 +95,27 @@ public class DocumentUploadIT extends IntegrationTest {
   }
 
   @Test
+  public void createReturnsErrorWhenFileIsEmpty() {
+    try {
+      PrintWriter writer = new PrintWriter("empty_file.png");
+      writer.close();
+    } catch (FileNotFoundException e) {
+      fail("Cannot open empty_file.png to write");
+    }
+
+    DocumentUploadRequest uploadRequest = new DocumentUploadRequest(DocumentUpload.Kind.EVIDENCE_DOCUMENT,
+        new File("empty_file.png"));
+
+    Result<DocumentUpload> uploadResult = gateway.documentUpload().create(uploadRequest);
+    ValidationError error = uploadResult.getErrors()
+      .forObject("documentUpload")
+      .getAllValidationErrors()
+      .get(0);
+
+    assertEquals(ValidationErrorCode.DOCUMENT_UPLOAD_FILE_IS_EMPTY, error.getCode());
+  }
+
+  @Test
   public void createReturnsErrorWhenPDFFileIsOver50Pages() {
     URL fileToUpload = getClass().getClassLoader().getResource("fixtures/too_long.pdf");
     DocumentUploadRequest uploadRequest = new DocumentUploadRequest(DocumentUpload.Kind.EVIDENCE_DOCUMENT,