+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## Unreleased
 * Add missing `paymentMethodToken` search to `CreditCardVerificationSearchRequest`
 * Add `recurringCustomerConsent` and `recurringMaxAmount` parameters to `authenticationInsightOptions` in `PaymentMethodNonce.create()`
-
+* Add `DOCUMENT_UPLOAD_FILE_IS_EMPTY` error code
 
 ## 2.108.0
 * Fix null pointer exception when 3DS lookup returns a `422`